feat: reorganise model and trpc router naming
