+++ src/app/dashboard/admin/@users/page.tsx
@@ -2,7 +2,7 @@ import { api } from "~/trpc/server"
 import UserTable from "./table"
 
 export default async function AdminUserTable() {
-  const users = await api.user.getAllAdmin.query()
+  const users = await api.admin.users.getAll.query()
 
   return <UserTable data={users} />
 }
