+++ src/app/profile/[id]/page.tsx
@@ -3,7 +3,7 @@ import { api } from "~/trpc/server"
 import ProfilePage from "./profile/page"
 
 const Profile = async ({ params: { id } }: { params: { id: string } }) => {
-  const currentUser = await api.user.getCurrent.query()
+  const currentUser = await api.users.getCurrent.query()
 
   if (currentUser) {
     return <ProfilePage currentUser={currentUser} id={id} />
