+++ src/server/api/routers/admin/analytics/get-user-count.ts
@@ -1,0 +1,18 @@
+import { sql, isNotNull } from "drizzle-orm"
+import { User } from "~/server/db/schema"
+import { adminProcedure } from "~/server/api/trpc"
+
+export const getUserCount = adminProcedure.query(async ({ ctx }) => {
+  const [[userCount], [memberCount]] = await Promise.all([
+    ctx.db.select({ count: sql<number>`count(*)`.mapWith(Number) }).from(User),
+    ctx.db
+      .select({ count: sql<number>`count(*)`.mapWith(Number) })
+      .from(User)
+      .where(isNotNull(User.role)),
+  ])
+
+  return {
+    users: userCount?.count ?? 0,
+    members: memberCount?.count ?? 0,
+  }
+})
