+++ src/server/api/routers/admin/users/update-role.ts
@@ -1,0 +1,26 @@
+import { TRPCError } from "@trpc/server"
+import { z } from "zod"
+import { eq } from "drizzle-orm"
+import { adminProcedure } from "~/server/api/trpc"
+import { User } from "~/server/db/schema"
+import { NAMED_ROLES } from "~/lib/constants"
+
+export const updateRole = adminProcedure
+  .input(
+    z.object({
+      id: z.string().min(2, {
+        message: "User ID is required",
+      }),
+      role: z.enum(NAMED_ROLES).nullable(),
+      paymentID: z.string().optional(),
+    }),
+  )
+  .mutation(async ({ ctx, input }) => {
+    const [user] = await ctx.db.update(User).set({ role: input.role }).where(eq(User.id, input.id)).returning()
+
+    if (!user) {
+      throw new TRPCError({ code: "NOT_FOUND", message: `User with id: ${input.id} does not exist` })
+    }
+
+    return user
+  })
