+++ src/server/db/types.ts
@@ -1,0 +1,5 @@
+import { type User } from "./schema"
+
+export type User = typeof User.$inferSelect
+
+export type Payment = typeof User.$inferSelect
