+++ src/server/services/square.ts
@@ -1,0 +1,7 @@
+import { env } from "~/env"
+import { Client, Environment } from "square"
+
+export const squareClient = new Client({
+  accessToken: env.SQUARE_ACCESS_TOKEN,
+  environment: env.NEXT_PUBLIC_SQUARE_APP_ID.includes("sandbox") ? Environment.Sandbox : Environment.Production,
+})
